+++ src/inputs/dropdownMenu/dropdownMenuConstants.js
@@ -3,3 +3,5 @@ const bemBlockName = 'dropdown_menu';
 export const BEM_BLOCK_NAME = bemBlockName;
 export const OPTION_CLASS_NAME = `${bemBlockName}--option`;
 export const OPTION_INNER_CLASS_NAME = `${bemBlockName}--option_inner`;
+export const OPTIONS_THEME_DARK = 'dark';
+export const OPTIONS_THEME_LIGHT = 'light';