+++ src/versions.js
@@ -1,97 +1,24 @@
 const versions = {
     'react-cm-ui': {
-        package: '9.14.1',
+        package: '9.16.0',
         components: {
-            dataDisplay: {
-                banner: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                divided: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                icon: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                image: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                list: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                personPanel: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                table: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
+            inputs: {
+                durationPicker: {
+                    devLibraryVersion: '2.0.0',
+                    designLibraryVersion: '', // TODO: Add Design Library version
+                    designLibraryDoc: '', // TODO: Add Sketch Cloud Link
                 },
             },
-            inputs: {
-                button: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
+            navigation: {
+                breadcrumbs: {
+                    devLibraryVersion: '2.0.0',
+                    designLibraryVersion: '', // TODO: Add Design Library version
+                    designLibraryDoc: '', // TODO: Add Sketch Cloud Link
                 },
-                checkbox: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                datePicker: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                datePickerCalendar: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                datePickerInput: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                dropdown: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                dropdownButton: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                durationPicker: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                input: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                phoneInput: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                prompt: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                radio: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                segmentedControls: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                textArea: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
-                },
-                timePicker: {
-                    devLibraryVersion: 'N/A',
-                    designLibraryVersion: 'N/A',
+                sectionalTabs: {
+                    devLibraryVersion: '2.0.0',
+                    designLibraryVersion: '', // TODO: Add Design Library version
+                    designLibraryDoc: '', // TODO: Add Sketch Cloud Link
                 },
             },
         },